Week 3 & 4: Finish basic sim + add eqns writeup
This also changes the units of each value to match with the "standard"
SI units for that value, because we'd otherwise have deal with weird
things like viscosity in terms of mega-grams, which... wasn't ideal.

But really the big thing here is the full simluation working.

Fully implement initial simulation

Add equations & units explainer, update unit docs in source