Week 5: now working! new CLI options for running
Turns out the bug was just a flipped sign for the volume in the
Jacobian -- now it works! (At least within reason.)

Added various CLI bits and pieces; now there's a CSV output option and
you can select between that & generating PNGs. The depth of the model
can also be selected as a command-line option.

Also: the way that we represent pressure & volume while doing a
simulation tick has been changed just a bit. They're now both relative
to atmospheric pressure (with volume being relative to what the *volume*
would be at atmospheric pressure). This gets the values MUCH closer in
magnitude to 1 so that the adding small values of dx to them is easier.