Add JSON model generator (`crate::gen::from_json`)
Generation reads from a supplied JSON config file, pre-generates & then
copies the `ChildInfo`s back in the implementation of `BranchGenerator`.

A sample model is provided as 'sample-model.json', which appears vaguely
similar to the human lungs, with the right bronchus (and all its
descendant bronchi) at 30% of the normal radius.